SGE - Pneuma's party-count setting no longer wipes the rest of the saved Sage settings on load

+++ Magitek/Models/Sage/SageSettings.cs
@@ -613,7 +613,7 @@ namespace Magitek.Models.Sage
 
         [Setting]
         [DefaultValue(3)]
-        public int PneumaNeedHealing { get; set; }
+        public float PneumaNeedHealing { get; set; }
 
         #endregion
